Initial split of stages into separate files
