+++ sto_sister/matching/ssim/common.py
