+++ sto_sister/pipeline/__init__.py
@@ -1,0 +1,4 @@
+from .core import Slot, Stage, StageResult, PipelineContext
+from .pipeline import SISTER, build_default_pipeline
+
+__all__ = ["Slot", "Stage", "StageResult", "PipelineContext", "SISTER", "build_default_pipeline"]
