+++ sto_sister/pipeline/pipeline.py
@@ -1,5 +1,5 @@
 from contextlib import contextmanager
-from dataclasses import dataclass, field
+#from dataclasses import dataclass, field
 from typing import Any, Callable, Dict, List, Tuple, Optional
 
 import time
@@ -8,293 +8,21 @@ import numpy as np
 from pathlib import Path
 
 # --- Import modules ---
-from src.exceptions import *
-
-from src.locator import LabelLocator
-from src.classifier import Classifier
-from src.region import RegionDetector
-from src.iconslot import IconSlotDetector
-from src.iconmatch import IconMatcher
-from src.prefilter import IconPrefilter
-from src.hashindex import HashIndex
-
-
-# --- Core value objects ---
-@dataclass(frozen=True)
-class Slot:
-    region_label: str
-    index: int
-    bbox: Tuple[int, int, int, int]
-
-
-@dataclass
-class PipelineContext:
-    screenshot: np.ndarray
-    config: Dict[str, Any] = field(default_factory=dict)
-    app_config: Dict[str, Any] = field(default_factory=dict)
-    labels: Dict[str, Tuple[int, int, int, int]] = field(default_factory=dict)
-    regions: Dict[str, Tuple[int, int, int, int]] = field(default_factory=dict)
-    slots: Dict[str, List[Slot]] = field(default_factory=dict)
-    classification: Dict[Slot, Any] = field(default_factory=dict)
-    predicted_qualities: Any = None
-    predicted_icons: Any = None
-    found_icons: Any = None
-    filtered_icons: Any = None
-
-
-@dataclass
-class StageResult:
-    """
-    Holds the context after a stage and any stage-specific output.
-    """
-
-    context: PipelineContext
-    output: Any
-
-
-# --- Abstract Stage ---
-class Stage:
-    name: str = ""
-    interactive: bool = False
-
-    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
-        self.opts = opts
-        self.app_config = app_config
-
-    def run(
-        self, ctx: PipelineContext, report: Callable[[str, float], None]
-    ) -> StageResult:
-        """
-        Execute the stage, updating ctx in place or returning a new one.
-        Use report(stage_name, percent_complete) to emit progress.
-        """
-        raise NotImplementedError
-
-
-# --- Concrete Stages ---
-class LabelLocatorStage(Stage):
-    name = "label_locator"
-
-    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
-        super().__init__(opts, app_config)
-        self.locator = LabelLocator(**opts)
-
-    def run(
-        self, ctx: PipelineContext, report: Callable[[str, float], None]
-    ) -> StageResult:
-        report(self.name, 0.0)
-        ctx.labels = self.locator.locate(ctx.screenshot)
-        report(self.name, 1.0)
-        return StageResult(ctx, ctx.labels)
-
-
-class ClassifierStage(Stage):
-    name = "classifier"
-
-    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
-        super().__init__(opts, app_config)
-        self.classifier = Classifier(**opts)
-
-    def run(
-        self, ctx: PipelineContext, report: Callable[[str, float], None]
-    ) -> StageResult:
-        report(self.name, 0.0)
-        ctx.classification = self.classifier.classify(ctx.labels)
-
-        if (
-            ctx.classification["build_type"] == "PC Ship Build"
-            or ctx.classification["build_type"] == "Console Ship Build"
-        ):
-            ctx.classification["icon_set"] = "ship"
-
-        elif ctx.classification["build_type"] == "PC Ground Build":
-            ctx.classification["icon_set"] = "pc_ground"
-
-        elif ctx.classification["build_type"] == "Console Ground Build":
-            ctx.classification["icon_set"] = "console_ground"
-
-        report(self.name, 1.0)
-        return StageResult(ctx, ctx.classification)
-
-
-class RegionDetectionStage(Stage):
-    name = "region_detection"
-    interactive = True  # allow UI confirmation
-
-    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
-        super().__init__(opts, app_config)
-        self.detector = RegionDetector(**opts)
-
-    def run(
-        self, ctx: PipelineContext, report: Callable[[str, float], None]
-    ) -> StageResult:
-        report(self.name, 0.0)
-
-        ctx.regions = self.detector.detect_regions(
-            ctx.screenshot, ctx.labels, ctx.classification
-        )
-
-        report(self.name, 1.0)
-        return StageResult(ctx, ctx.regions)
-
-
-class IconSlotDetectionStage(Stage):
-    name = "iconslot_detection"
-
-    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
-        super().__init__(opts, app_config)
-
-        self.opts["hash_index"] = self.app_config.get("hash_index")
-
-        self.slot_detector = IconSlotDetector(**self.opts)
-
-    def run(
-        self, ctx: PipelineContext, report: Callable[[str, float], None]
-    ) -> StageResult:
-        report(self.name, 0.0)
-
-        ctx.slots = self.slot_detector.detect_slots(ctx.screenshot, ctx.regions)
-
-        report(self.name, 1.0)
-        return StageResult(ctx, ctx.slots)
-
-
-class IconPrefilterStage(Stage):
-    name = "icon_prefilter"
-
-    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
-        super().__init__(opts, app_config)
-
-        self.opts["hash_index"] = self.app_config.get("hash_index")
-
-        self.prefilterer = IconPrefilter(**self.opts)
-
-    def run(
-        self, ctx: PipelineContext, report: Callable[[str, float], None]
-    ) -> StageResult:
-        report(self.name, 0.0)
-
-        icon_sets = ctx.app_config.get("icon_sets", {})
-        icon_set = icon_sets[ctx.classification["icon_set"]]
-
-        ctx.predicted_icons = self.prefilterer.icon_predictions(ctx.slots, icon_set)
-        ctx.found_icons = self.prefilterer.found_icons
-        ctx.filtered_icons = self.prefilterer.filtered_icons
-
-        report(self.name, 1.0)
-        return StageResult(ctx, ctx.predicted_icons)
-
-
-class IconMatchingQualityDetectionStage(Stage):
-    name = "icon_quality_detection"
-
-    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
-        super().__init__(opts, app_config)
-
-        self.matcher = IconMatcher(
-            hash_index=app_config.get("hash_index"),
-            debug=opts.get("debug", False),
-            engine_type=opts.get("engine_type", "ssim"),
-        )
-
-    def run(
-        self, ctx: PipelineContext, report: Callable[[str, float], None]
-    ) -> StageResult:
-        report(self.name, 0.0)
-
-        overlays = self.matcher.load_quality_overlays(ctx.config.get("overlay_dir", ""))
-
-        ctx.predicted_qualities = self.matcher.quality_predictions(
-            ctx.slots,
-            overlays,
-            threshold=self.opts.get("threshold", 0.8),
-        )
-        report(self.name, 1.0)
-        return StageResult(ctx, ctx.predicted_qualities)
-
-
-class IconMatchingStage(Stage):
-    name = "icon_matching"
-
-    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
-        super().__init__(opts, app_config)
-
-        self.matcher = IconMatcher(
-            hash_index=app_config.get("hash_index"),
-            debug=opts.get("debug", False),
-            engine_type=opts.get("engine_type", "ssim"),
-        )
-
-    def run(
-        self, ctx: PipelineContext, report: Callable[[str, float], None]
-    ) -> StageResult:
-        report(self.name, 0.0)
-        icon_sets = ctx.app_config.get("icon_sets", {})
-        ctx.overlays = self.matcher.load_quality_overlays(
-            ctx.config.get("overlay_dir", "")
-        )
-        # print(f"[Matching] ctx.filtered_icons: {ctx.filtered_icons}")
-        ctx.matches = self.matcher.match_all(
-            ctx.slots,
-            icon_sets,
-            ctx.overlays,
-            ctx.predicted_qualities,
-            ctx.filtered_icons,
-            ctx.found_icons,
-            threshold=self.opts.get("threshold", 0.7),
-        )
-        report(self.name, 1.0)
-        return StageResult(ctx, ctx.matches)
-
-
-class OutputTransformationStage(Stage):
-    name = "output_transformation"
-
-    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
-        super().__init__(opts, app_config)
-
-    def run(
-        self, ctx: PipelineContext, report: Callable[[str, float], None]
-    ) -> StageResult:
-        report(self.name, 0.0)
-
-        # start with whatever matches we already have
-        ctx.output = {
-            "matches": ctx.matches,
-            "predicted_icons": ctx.predicted_icons,
-            "predicted_qualities": ctx.predicted_qualities,
-        }
-
-        # we're going to merge any predicted icons into ctx.matches if they don't already exist
-        # this catches cases where we have predicted icons but no matches, so we at least provide some output that is hopefully useful
-        matches = ctx.matches
-        for region in ctx.predicted_icons:
-            region_name = region
-            for slot in ctx.predicted_icons[region]:
-                slot_name = slot
-
-                # find any predicted icons for this region/slot
-                predicted = ctx.predicted_icons.get(region_name, {}).get(slot_name, [])
-
-                # check current output for this region/slot
-                # print(f"region_name: {region_name}, slot_name: {slot_name} existing: {matches.get(region_name, {}).get(slot_name, [])}")
-                existing = matches.get(region_name, {}).get(slot_name, [])
-
-                # if there's a prediction and no existing match, copy it over
-                if predicted and len(existing) == 0:
-                    # ensure the dicts exist
-                    matches.setdefault(region_name, {})[slot_name] = predicted.copy()
-
-                    # copy over the predicted quality if we have it
-                    if ctx.predicted_qualities[region_name]:
-                        for idx, item in enumerate(matches[region_name][slot_name]):
-                            matches[region_name][slot_name][idx][
-                                "predicted_quality"
-                            ] = ctx.predicted_qualities[region_name][slot_name]
-
-        report(self.name, 1.0)
-        return StageResult(ctx, ctx.output)
-
+from .core import *
+from ..exceptions import *
+
+from ..hashindex import HashIndex
+
+from ..stages import (
+    LabelLocatorStage,
+    ClassifierStage,
+    RegionDetectionStage,
+    IconSlotDetectionStage,
+    IconPrefilterStage,
+    IconMatchingQualityDetectionStage,
+    IconMatchingStage,
+    OutputTransformationStage,
+)   
 
 # --- The Pipeline Orchestrator ---
 class SISTER:
