+++ sto_sister/stages/classifier.py
@@ -1,0 +1,32 @@
+from typing import Any, Callable, Dict, List, Tuple, Optional
+
+from ..pipeline import Stage, StageResult, PipelineContext
+from ..classifier import Classifier
+
+class ClassifierStage(Stage):
+    name = "classifier"
+
+    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
+        super().__init__(opts, app_config)
+        self.classifier = Classifier(**opts)
+
+    def run(
+        self, ctx: PipelineContext, report: Callable[[str, float], None]
+    ) -> StageResult:
+        report(self.name, 0.0)
+        ctx.classification = self.classifier.classify(ctx.labels)
+
+        if (
+            ctx.classification["build_type"] == "PC Ship Build"
+            or ctx.classification["build_type"] == "Console Ship Build"
+        ):
+            ctx.classification["icon_set"] = "ship"
+
+        elif ctx.classification["build_type"] == "PC Ground Build":
+            ctx.classification["icon_set"] = "pc_ground"
+
+        elif ctx.classification["build_type"] == "Console Ground Build":
+            ctx.classification["icon_set"] = "console_ground"
+
+        report(self.name, 1.0)
+        return StageResult(ctx, ctx.classification)
