+++ sto_sister/stages/label_locator.py
@@ -1,0 +1,19 @@
+from typing import Any, Callable, Dict, List, Tuple, Optional
+
+from ..pipeline import Stage, StageResult, PipelineContext
+from ..locator import LabelLocator
+
+class LabelLocatorStage(Stage):
+    name = "label_locator"
+
+    def __init__(self, opts: Dict[str, Any], app_config: Dict[str, Any]):
+        super().__init__(opts, app_config)
+        self.locator = LabelLocator(**opts)
+
+    def run(
+        self, ctx: PipelineContext, report: Callable[[str, float], None]
+    ) -> StageResult:
+        report(self.name, 0.0)
+        ctx.labels = self.locator.locate(ctx.screenshot)
+        report(self.name, 1.0)
+        return StageResult(ctx, ctx.labels)
