+++ sto_sister/utils/image.py
